kolejne kroki

+++ Project/STM8S_StdPeriph_Template/BMP.h
@@ -3,7 +3,7 @@
 
 void BMP_init(bmp_t * BMP);
 void BMP_read(bmp_t * BMP);
-float BMP_altitude(float startPress, float currPress);
+int32_t BMP_altitude(int32_t startPress, int32_t currPress);
 
 
 #endif
